Commenting out non movement code
I am not feeling well and am unsure  if I am going to make it to the meeting so I am pushing the commented out code. Jay should look over this to make sure I commented out everything that needed to be. The only non comment edits I made were to the  imports of swervemodule and drivetrain since they are both no longer in a subsystems folder.